requirements.txt
annotated-types==0.7.0
anyio==4.4.0
api-display-purposes==0.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
clarifai==10.5.3
clarifai-grpc==10.5.4
colorama==0.4.6
configparser==3.8.1
contextlib2==21.6.0
decorator==4.4.2
EasyProcess==1.1
emoji==1.6.3
exceptiongroup==1.2.1
future==1.0.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huepy==1.2.1
idna==2.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
inquirerpy==0.3.4
insta-api==0.2.4
instabot==0.117.0
instagrapi==2.1.2
instapy==0.6.16
jsonschema==2.6.0
log3==0.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
MeaningCloud-python==2.0.0
mock==5.1.0
moviepy==1.0.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pfzy==0.3.4
pillow==10.3.0
pluggy==1.5.0
plyer==2.1.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==3.20.3
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PySocks==1.7.1
pytest==8.2.2
python-rapidjson==1.17
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.3
rich==13.7.1
schedule==1.2.2
schema==0.7.5
selenium==4.22.0
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabulate==0.9.0
tomli==2.0.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
tritonclient==2.47.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webdriverdownloader==1.1.0.4
websocket-client==1.8.0
wsproto==1.2.0